requirements.txt
aiofiles>=23.2.1
aiohttp>=3.8.5
chardet>=5.2.0
deepgram-sdk>=2.3.0
fastapi>=0.104.1
google-generativeai>=0.3.2
httptools>=0.6.3
langchain==0.1.17
langchain-core==0.1.52
langchain-google-genai==0.0.6
langgraph==0.0.41
numpy>=1.24.0,<2.0.0
python-dotenv>=1.0.0
python-multipart>=0.0.6
requests>=2.31.0
uvicorn>=0.24.0
websockets>=12.0